4.8

+++ src/main/java/com/example/examsystem/common/BaseContext.java
@@ -6,8 +6,8 @@ package com.example.examsystem.common;
 public class BaseContext {
     private static ThreadLocal<Long> threadLocal=new ThreadLocal<>();
 
-    public static void setCurrentId(Long id){
-        threadLocal.set(id);
+    public static void setCurrentId(Integer id){
+        threadLocal.set(Long.valueOf(id));
     }
 
     public static Long getCurrentId(){
